fix(setup): Fixed Git conflict
