+++ .idea/jarRepositories.xml
@@ -9,61 +9,52 @@
     <remote-repository>
       <option name="id" value="jboss.community" />
       <option name="name" value="JBoss Community repository" />
-      <option name="url"
-        value="https://repository.jboss.org/nexus/content/repositories/public/" />
+      <option name="url" value="https://repository.jboss.org/nexus/content/repositories/public/" />
     </remote-repository>
     <remote-repository>
       <option name="id" value="WPILocal" />
       <option name="name" value="WPILocal" />
-      <option name="url"
-        value="file:/$USER_HOME$/Public/wpilib/2023/maven/" />
+      <option name="url" value="file:/$USER_HOME$/Public/wpilib/2023/maven/" />
     </remote-repository>
     <remote-repository>
       <option name="id" value="MavenRepo" />
       <option name="name" value="MavenRepo" />
-      <option name="url"
-        value="https://repo.maven.apache.org/maven2/" />
+      <option name="url" value="https://repo.maven.apache.org/maven2/" />
     </remote-repository>
     <remote-repository>
       <option name="id" value="WPIOfficialRelease" />
       <option name="name" value="WPIOfficialRelease" />
-      <option name="url"
-        value="https://frcmaven.wpi.edu/artifactory/release" />
+      <option name="url" value="https://frcmaven.wpi.edu/artifactory/release" />
     </remote-repository>
     <remote-repository>
       <option name="id" value="WPILocal" />
       <option name="name" value="WPILocal" />
-      <option name="url"
-        value="file:/$PROJECT_DIR$/../../../../../../Public/wpilib/2023/maven/" />
+      <option name="url" value="file:/$PROJECT_DIR$/../../../../../../Public/wpilib/2023/maven/" />
     </remote-repository>
     <remote-repository>
       <option name="id" value="WPIFRCMavenVendorCacheRelease" />
       <option name="name" value="WPIFRCMavenVendorCacheRelease" />
-      <option name="url"
-        value="https://frcmaven.wpi.edu/artifactory/vendor-mvn-release" />
+      <option name="url" value="https://frcmaven.wpi.edu/artifactory/vendor-mvn-release" />
     </remote-repository>
     <remote-repository>
-      <option name="id"
-        value="WPI3f48eb8c-50fe-43a6-9cb7-44c86353c4cb_0Release" />
-      <option name="name"
-        value="WPI3f48eb8c-50fe-43a6-9cb7-44c86353c4cb_0Release" />
+      <option name="id" value="WPI3f48eb8c-50fe-43a6-9cb7-44c86353c4cb_0Release" />
+      <option name="name" value="WPI3f48eb8c-50fe-43a6-9cb7-44c86353c4cb_0Release" />
       <option name="url" value="https://maven.revrobotics.com/" />
     </remote-repository>
     <remote-repository>
-      <option name="id"
-        value="WPIcb311d09-36e9-4143-a032-55bb2b94443b_0Release" />
-      <option name="name"
-        value="WPIcb311d09-36e9-4143-a032-55bb2b94443b_0Release" />
-      <option name="url"
-        value="https://dev.studica.com/maven/release/2023/" />
+      <option name="id" value="WPIcb311d09-36e9-4143-a032-55bb2b94443b_0Release" />
+      <option name="name" value="WPIcb311d09-36e9-4143-a032-55bb2b94443b_0Release" />
+      <option name="url" value="https://dev.studica.com/maven/release/2023/" />
     </remote-repository>
     <remote-repository>
-      <option name="id"
-        value="WPIab676553-b602-441f-a38d-f1296eff6537_0Release" />
-      <option name="name"
-        value="WPIab676553-b602-441f-a38d-f1296eff6537_0Release" />
-      <option name="url"
-        value="https://maven.ctr-electronics.com/release/" />
+      <option name="id" value="WPIab676553-b602-441f-a38d-f1296eff6537_0Release" />
+      <option name="name" value="WPIab676553-b602-441f-a38d-f1296eff6537_0Release" />
+      <option name="url" value="https://maven.ctr-electronics.com/release/" />
+    </remote-repository>
+    <remote-repository>
+      <option name="id" value="WPILocal" />
+      <option name="name" value="WPILocal" />
+      <option name="url" value="file:/$PROJECT_DIR$/../../../../Public/wpilib/2023/maven/" />
     </remote-repository>
   </component>
 </project>
