+++ .idea/misc.xml
@@ -1,10 +1,7 @@
 <project version="4">
-  <component name="ExternalStorageConfigurationManager"
-    enabled="true" />
+  <component name="ExternalStorageConfigurationManager" enabled="true" />
   <component name="FrameworkDetectionExcludesConfiguration">
     <file type="web" url="file://$PROJECT_DIR$" />
   </component>
-  <component name="ProjectRootManager" version="2"
-    languageLevel="JDK_11" project-jdk-name="17"
-    project-jdk-type="JavaSDK" />
+  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" project-jdk-name="17" project-jdk-type="JavaSDK" />
 </project>
