+++ src/main/java/org/robolancers321/RobotContainer.java
@@ -1,39 +1,67 @@
 /* (C) Robolancers 2024 */
 package org.robolancers321;
 
-import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
+import static org.robolancers321.Constants.Swerve.*;
+import static org.robolancers321.Constants.OperatorConstants.*;
+
+import com.kauailabs.navx.frc.AHRS;
+
+import edu.wpi.first.math.MathUtil;
+import edu.wpi.first.wpilibj.SPI;
+import edu.wpi.first.wpilibj.smartdashboard.Field2d;
 import edu.wpi.first.wpilibj2.command.Command;
+import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
-import java.io.FileNotFoundException;
+import org.robolancers321.Constants;
 import org.robolancers321.subsystems.arm.Arm;
 import org.robolancers321.subsystems.arm.commands.RunArm;
-import org.robolancers321.util.VersionLoader;
+import org.robolancers321.subsystems.swerve.Swerve;
+import org.robolancers321.subsystems.swerve.SwerveModule;
 
 public class RobotContainer {
-  private Arm arm = new Arm();
+  private final Field2d field = new Field2d();
+  private final CommandXboxController driver = new CommandXboxController(Constants.OperatorConstants.kDriverControllerPort);
+  private final AHRS gyro = new AHRS(SPI.Port.kMXP);
+  private final Arm arm = new Arm();
 
-  private final CommandXboxController controller =
-      new CommandXboxController(Constants.OperatorConstants.kDriverControllerPort);
+
+  private final Swerve swerve =
+      new Swerve(
+          new SwerveModule(frontLeft),
+          new SwerveModule(frontRight),
+          new SwerveModule(backLeft),
+          new SwerveModule(backRight),
+          gyro,
+          field);
 
   public RobotContainer() {
+    swerve.setDefaultCommand(
+      swerve.drive(
+        this::getThrottle,
+        this::getStrafe,
+        this::getTurn,
+        true));
+
     this.arm.setDefaultCommand(new RunArm(arm));
 
     configureBindings();
-
-    try {
-      String versionData = new VersionLoader().getVersionData().toString();
-      System.out.println(versionData);
-      SmartDashboard.putString("VERSION_DATA", versionData);
-
-    } catch (FileNotFoundException ex) {
-      System.out.println("VERSIONING FILE NOT FOUND");
-      SmartDashboard.putString("VERSION_DATA", "VERSIONING FILE NOT FOUND");
-    }
   }
 
   private void configureBindings() {}
 
   public Command getAutonomousCommand() {
-    return null;
+    return Commands.print("No autonomous command configured");
+  }
+
+  private double getThrottle() {
+    return kMaxSpeedMetersPerSecond * MathUtil.applyDeadband(driver.getLeftY(), kJoystickDeadband);
+  }
+
+  private double getStrafe() {
+    return kMaxSpeedMetersPerSecond * MathUtil.applyDeadband(driver.getLeftX(), kJoystickDeadband);
+  }
+
+  private double getTurn() {
+    return kMaxOmegaRadiansPerSecond * MathUtil.applyDeadband(driver.getRightX(), kJoystickDeadband);
   }
 }
