+++ src/main/java/org/robolancers321/commands/autos/TaxiAndScore.java
@@ -1,0 +1,4 @@
+package org.robolancers321.commands.autos;
+
+public class TaxiAndScore {
+}
