+++ src/main/java/org/robolancers321/subsystems/arm/commands/RunArm.java
@@ -2,7 +2,6 @@
 package org.robolancers321.subsystems.arm.commands;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import org.robolancers321.Constants;
 import org.robolancers321.subsystems.arm.Arm;
 
 public class RunArm extends CommandBase {
